output changed to files instead of console / relocated code to cli_conda_merge.py
