+++ cli_code/conda_merge/Input YAMLS/another-environment.yml
@@ -1,0 +1,7 @@
+channels:
+  - conda
+  - conda-forge
+dependencies:
+  - celery==3.1.25
+  - django-celery
+  - fabric
