+++ cli_code/conda_merge/Input YAMLS/dev-environment.yml
@@ -1,0 +1,6 @@
+channels:
+  - conda
+dependencies:
+  - fabric
+  - pip:
+    - genty
