+++ cli_code/conda_merge/Input YAMLS/my-environment.yml
@@ -1,0 +1,10 @@
+channels:
+  - defaults
+  - conda
+  - conda-forge
+dependencies:
+  - jupyter
+  - pudb
+  - pylint
+  - backports.functools_lru_cache # for pylint
+  - isort # for pylint
