--- merge_conda_yaml/.gitignore
@@ -1,6 +1,0 @@
-*.pyc
-.cache
-build
-dist
-*.egg-info
-refer
